Added slp.openWallet() method and unit test
